+++ Tests/Loader/XmlFileLoaderTest.php
@@ -1335,19 +1335,6 @@ class XmlFileLoaderTest extends TestCase
         $loader->load('key_type_wrong_constant.xml');
     }
 
-    /**
-     * @group legacy
-     */
-    public function testDeprecatedTagged()
-    {
-        $container = new ContainerBuilder();
-        $loader = new XmlFileLoader($container, new FileLocator(self::$fixturesPath.'/xml'));
-
-        $this->expectUserDeprecationMessage(\sprintf('Since symfony/dependency-injection 7.2: Type "tagged" is deprecated for tag <argument>, use "tagged_iterator" instead in "%s/xml%sservices_with_deprecated_tagged.xml".', self::$fixturesPath, \DIRECTORY_SEPARATOR));
-
-        $loader->load('services_with_deprecated_tagged.xml');
-    }
-
     public function testLoadServicesWithEnvironment()
     {
         $container = new ContainerBuilder();